Revert 7d9f4f6..6f495ab
This rolls back to commit 7d9f4f6f2751af2b6724851fdff52a37c877ef29.

+++ TimeColor/AppDelegate.m
@@ -192,7 +192,7 @@ static void DrawGradientPattern(void * info, CGContextRef context)
 @synthesize launchAtLoginButton;
 -(IBAction)toggleLaunchAtLogin:(id)sender
 {
-    NSInteger clickedSegment = [sender selectedSegment];
+    int clickedSegment = [sender selectedSegment];
     if (clickedSegment == 0) { // ON
         // Turn on launch at login
         if (!SMLoginItemSetEnabled ((__bridge CFStringRef)@"com.issuepress.TimeColorHelper", YES)) {
